+++ lib/intercept.js
@@ -281,6 +281,8 @@ function overrideClientRequest() {
       debug('using', interceptors.length, 'interceptors')
 
       //  Use filtered interceptors to intercept requests.
+      // TODO: this shouldn't be a class anymore
+      // the overrider explicitly overrides methods and attrs on the request so the `assign` below should be removed.
       const overrider = new InterceptedRequestRouter({
         req: this,
         options,
