+++ lib/interceptor.js
@@ -73,7 +73,6 @@ module.exports = class Interceptor {
 
     this.delayBodyInMs = 0
     this.delayConnectionInMs = 0
-    this.delaySocketInMs = 0
 
     this.optional = false
 
@@ -599,7 +598,7 @@ module.exports = class Interceptor {
    * @return {Interceptor} - the current interceptor for chaining
    */
   delayBody(ms) {
-    this.delayBodyInMs += ms
+    this.delayBodyInMs = ms
     return this
   }
 
@@ -610,26 +609,7 @@ module.exports = class Interceptor {
    * @return {Interceptor} - the current interceptor for chaining
    */
   delayConnection(ms) {
-    this.delayConnectionInMs += ms
-    return this
-  }
-
-  /**
-   * @private
-   * @returns {number}
-   */
-  getTotalDelay() {
-    return this.delayBodyInMs + this.delayConnectionInMs
-  }
-
-  /**
-   * Make the socket idle for a certain number of ms (simulated).
-   *
-   * @param  {integer} ms - Number of milliseconds to wait
-   * @return {Interceptor} - the current interceptor for chaining
-   */
-  socketDelay(ms) {
-    this.delaySocketInMs = ms
+    this.delayConnectionInMs = ms
     return this
   }
 }
