+++ lib/playback_interceptor.js
@@ -137,15 +137,15 @@ function playbackInterceptor({
     interceptor.scope.emit('request', req, interceptor, requestBodyString)
 
     if (typeof interceptor.errorMessage !== 'undefined') {
-      interceptor.markConsumed()
-
       let error
       if (typeof interceptor.errorMessage === 'object') {
         error = interceptor.errorMessage
       } else {
         error = new Error(interceptor.errorMessage)
       }
-      common.setTimeout(() => emitError(error), interceptor.getTotalDelay())
+
+      const delay = interceptor.delayBodyInMs + interceptor.delayConnectionInMs
+      common.setTimeout(emitError, delay, error)
       return
     }
 
@@ -291,7 +291,7 @@ function playbackInterceptor({
       response.emit('error', err)
     })
 
-    const { delayBodyInMs, delayConnectionInMs, delaySocketInMs } = interceptor
+    const { delayBodyInMs, delayConnectionInMs } = interceptor
 
     function respond() {
       if (req.aborted) {
@@ -310,7 +310,7 @@ function playbackInterceptor({
       common.setTimeout(() => bodyAsStream.resume(), delayBodyInMs)
     }
 
-    socket.applyDelay(delaySocketInMs + delayConnectionInMs)
+    socket.applyDelay(delayConnectionInMs)
     common.setTimeout(respond, delayConnectionInMs)
   }
